occultatum-python/pyproject.toml
[project]
name = "occultatum"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy==2.0.0",
    "onnxruntime>=1.22.1",
    "opencv-contrib-python>=4.12.0.88",
    "pillow>=11.3.0",
    "rembg>=2.0.67",
]

[dependency-groups]
dev = [
    "polylith-cli>=1.30.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
dev-mode-dirs = ["components", "bases", "development", "."]

[tool.uv.workspace]
members = [
    "development",
]

[tool.pyright]
pythonVersion = "3.11"
include = ["components", "bases"]

[tool.polylith.bricks]
"bases/occultatum/artifact_to_3d_cli" = "occultatum/artifact_to_3d_cli"
"components/occultatum/artifact_to_3d" = "occultatum/artifact_to_3d"
"components/occultatum/crop_image_to_foreground" = "occultatum/crop_image_to_foreground"
"components/occultatum/isolate_artifact" = "occultatum/isolate_artifact"
